+++ src/interpreter/LuaManager.cpp
@@ -1,4 +1,5 @@
 #include "LuaManager.hpp"
+#include "LuaObject.hpp"
 
 using namespace geode::prelude;
 
@@ -96,104 +97,6 @@ std::string LuaManager::highlightSyntax(const std::string& code) {
         i++;
     }
     return result;
-}
-
-void moveGroupWithEasing(
-    GJBaseGameLayer* gameLayer, 
-    int targetGroupID, 
-    cocos2d::CCPoint offset, 
-    float duration, 
-    int easingType = 0, 
-    float easingRate = 2.0f
-) {
-    if (!gameLayer) return;
-
-    auto moveTrigger = static_cast<EffectGameObject*>(GameObject::createWithKey(901));
-    if (!moveTrigger) {
-        log::error("Move trigger cast failed");
-        return;
-    }
-
-    moveTrigger->m_targetGroupID = targetGroupID;
-    moveTrigger->m_moveOffset = offset;
-    moveTrigger->m_duration = duration;
-    moveTrigger->m_easingType = static_cast<EasingType>(easingType);
-    moveTrigger->m_easingRate = easingRate;
-
-    moveTrigger->triggerObject(gameLayer, -1, nullptr);
-}
-
-void rotateGroupWithEasing(
-    GJBaseGameLayer* gameLayer, 
-    int targetGroupID, 
-    int centerGroupID, 
-    int degrees, 
-    int times360, 
-    float duration, 
-    int easingType = 0, 
-    float easingRate = 2.0f,
-    bool lockObjRotation = false
-) {
-    if (!gameLayer) return;
-
-    auto rotateTrigger = static_cast<EffectGameObject*>(GameObject::createWithKey(1346));
-    if (!rotateTrigger) {
-        log::error("Rotate trigger creation failed");
-        return;
-    }
-
-    rotateTrigger->m_targetGroupID = targetGroupID;
-    rotateTrigger->m_centerGroupID = centerGroupID;
-    rotateTrigger->m_rotationDegrees = degrees;
-    rotateTrigger->m_times360 = times360;
-    rotateTrigger->m_duration = duration;
-    rotateTrigger->m_easingType = static_cast<EasingType>(easingType);
-    rotateTrigger->m_easingRate = easingRate;
-    rotateTrigger->m_lockObjectRotation = lockObjRotation;
-
-    rotateTrigger->triggerObject(gameLayer, -1, nullptr);
-}
-
-void scaleGroupWithEasing(
-    GJBaseGameLayer* gameLayer, 
-    int targetGroupID, 
-    int centerGroupID, 
-    float scaleX, 
-    float scaleY, 
-    float duration, 
-    int easingType = 0, 
-    float easingRate = 2.0f,
-    bool divByX = false,
-    bool divByY = false,
-    bool onlyMove = false,
-    bool relativeScale = false,
-    bool relativeRotation = false
-) {
-    if (!gameLayer) return;
-
-    auto scaleTrigger = static_cast<TransformTriggerGameObject*>(GameObject::createWithKey(2067));
-    if (!scaleTrigger) {
-        log::error("Scale trigger creation failed");
-        return;
-    }
-
-    scaleTrigger->m_targetGroupID = targetGroupID;
-    scaleTrigger->m_centerGroupID = centerGroupID;
-
-    scaleTrigger->m_scaleX = scaleX;
-    scaleTrigger->m_scaleY = scaleY;
-    scaleTrigger->m_duration = duration;
-
-    scaleTrigger->m_easingType = static_cast<EasingType>(easingType);
-    scaleTrigger->m_easingRate = easingRate;
-
-    scaleTrigger->m_divideX = divByX;
-    scaleTrigger->m_divideY = divByY;
-    scaleTrigger->m_onlyMove = onlyMove;
-    scaleTrigger->m_relativeScale = relativeScale;
-    scaleTrigger->m_relativeRotation = relativeRotation;
-
-    scaleTrigger->triggerObject(gameLayer, -1, nullptr);
 }
 
 LuaManager::LuaManager() {
@@ -318,6 +221,51 @@ LuaManager::LuaManager() {
             relativeRotation
         );
     };
+    // Not ready yet!
+    /*m_lua->new_usertype<LuaObject>(
+        "LuaObject",
+
+        "addToLayer",
+        &LuaObject::addToLayer,
+
+        "setPosition",
+        &LuaObject::setPosition,
+
+        "getPosition",
+        &LuaObject::getPosition,
+
+        "move",
+        &LuaObject::move,
+
+
+        "setRotation",
+        &LuaObject::setRotation,
+
+        "getRotation",
+        &LuaObject::getRotation,
+
+        "rotate",
+        &LuaObject::rotate,
+
+
+        "setScale",
+        &LuaObject::setScale,
+
+        "getScale",
+        &LuaObject::getScale,
+
+        "scale",
+        &LuaObject::scale,
+
+
+        "getID",
+        &LuaObject::getID
+    );
+    groupTable["new"] = [layer](int id) -> LuaObject* {
+        auto obj = new LuaObject(layer,id);
+        obj->addToLayer();
+        return obj;
+    };*/
 
     m_lua->set_function("wait", [](double seconds, sol::this_state s) {
         lua_pushnumber(s, seconds);