+++ changelog.md
@@ -1,4 +1,9 @@
 # v1.0.0-alpha.1
 
 - Added `Player.kill()` function
-- Added `Group.move()` function
+- Added `Object.move()` function
+- Added `Object.rotate()` function
+- Added `Object.scale()` function
+- Added `Object.new()` function
+- Added `wait()` function
+- Added notifications to `print()`, `warn()`, and `error()` functions